React compiler linting, list fixes
