+++ eslint.config.mjs
@@ -6,6 +6,8 @@ import js from '@eslint/js';
 import { FlatCompat } from '@eslint/eslintrc';
 import andrewaylett from 'eslint-config-andrewaylett';
 import tseslint from 'typescript-eslint';
+import react from 'eslint-plugin-react';
+import reactCompiler from 'eslint-plugin-react-compiler';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -39,6 +41,11 @@ export default tseslint.config(
     ...andrewaylett,
     ...tseslint.configs.stylisticTypeChecked,
     {
+        ...react.configs.flat.recommended,
+        settings: { react: { version: 'detect' } },
+    },
+    reactCompiler.configs.recommended,
+    {
         rules: {
             '@typescript-eslint/restrict-template-expressions': [
                 'error',
@@ -47,10 +54,29 @@ export default tseslint.config(
                     allowBoolean: true,
                 },
             ],
+            'react/no-unescaped-entities': [
+                'error',
+                {
+                    forbid: [
+                        {
+                            char: '>',
+                            alternatives: ['&gt;'],
+                        },
+                        {
+                            char: '}',
+                            alternatives: ['&#125;'],
+                        },
+                    ],
+                },
+            ],
+            'react/no-unknown-property': [
+                'error',
+                { ignore: ['property', 'resource', 'typeof', 'vocab'] },
+            ],
         },
     },
     {
-        files: ['*.js', '*.mjs'],
+        files: ['*.js', '*.mjs', '*.ts'],
         languageOptions: {
             globals: {
                 ...globals.node,
