+++ src/app/styles/global.css
@@ -43,10 +43,10 @@
 
 @utility bg-draft {
     @apply fixed -z-10 rotate-45 fill-slate-700;
-    --tw-translate-x: -25%;
-    --tw-translate-y: -25%;
-    width: calc(100vw + 100vh);
-    height: calc(100vw + 100vh);
+    left: -50%;
+    top: -50%;
+    width: calc(200vw + 200vh);
+    height: calc(200vw + 200vh);
     /* fill=slate-200 */
     @apply bg-[url('data:image/svg+xml;charset=UTF-8,%3csvg%20width="60"%20height="50"%20viewBox="0%200%2050%2050"%20xmlns="http://www.w3.org/2000/svg"%3e%3ctext%20fill="oklch(0.929%200.013%20255.508)"%20x="0"%20y="50"%3eDRAFT%3c/text%3e%3c/svg%3e')];
     /* fill=slate-700 */
@@ -93,15 +93,15 @@
     }
 
     ol {
-        @apply list-decimal;
+        @apply list-decimal pl-[2ch];
     }
 
     ul {
-        @apply list-disc;
+        @apply list-disc pl-[2ch];
     }
 
     li {
-        @apply list-item list-inside;
+        @apply list-item list-outside;
     }
 
     nav {
