+++ tsconfig.json
@@ -36,6 +36,7 @@
   ],
   "include": [
     "next-env.d.ts",
+    "next.config.ts",
     "*/**/*.ts",
     "*/**/*.tsx",
     "*/**/*.mdx",
